fix: issue with handlers on docker
