+++ CMakeLists.txt
@@ -6,7 +6,7 @@ include_directories(../asabru-commons/src/config)
 include_directories(../libuv/include)
 
 # Set C++ standard
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
