+++ src/interface/CProxyHandler.h
@@ -1,7 +1,9 @@
 #ifndef PROXY_HANDLER_H
 #define PROXY_HANDLER_H
 
-#include "../socket/ServerSocket.h"
+#include <string>
+#include "CommonTypes.h"
+#include "../socket/CServerSocket.h"
 #include "../socket/SocketClient.h"
 
 class CProxyHandler
@@ -9,9 +11,9 @@ class CProxyHandler
 public:
     CProxyHandler() {}
 
-    virtual void * HandleUpstreamData(void *buffer, int buffer_length, SocketClient *target) = 0;
+    virtual std::string HandleUpstreamData(void *buffer, int buffer_length, EXECUTION_CONTEXT *exec_context) = 0;
 
-    virtual void * HandleDownStreamData(void *buffer, int buffer_length, Socket *client) = 0;
+    virtual std::string HandleDownStreamData(void *buffer, int buffer_length, EXECUTION_CONTEXT *exec_context) = 0;
 
     virtual ~CProxyHandler() {}
 };
