+++ src/socket/CServerSocket.h
@@ -16,6 +16,7 @@
 #endif
 
 #include "Socket.h"
+#include "SocketServer.h"
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
